feat: chat history #26
The chat history works by keeping track of the current chat we are looking at. When you click one of the chat history #s that are not the one you are currently on it will rebuild the chat, and run main() to prompt you to continue. Otherwise, if the branch you clicked on has no localStorage data, then it will regularly prompt you. One flaw that I found is the overallFortune variable crosses over from one history to the next. Also, I removed the functionality that you can delete one of the lines because for now it would be hard to track, and by removing the functionality we have the option to always update since the user can keep describing one palm line more than once.