feat: chat history #26
I have updated the code back to using dates in localStorage. Each button is loaded with appropriate key value correlating to it when the site is reloaded. Each button is able to switch to a different chat. The bugs right now are that if you are not done with one chat, and you switch to the next, then the main() functions kind of run at the same time and mess each other up. Also, the overallFortune is still adding to each other if you switch sessions.